requirements.txt
Flask
flask_sqlalchemy
flask_bcrypt
flask_jwt_extended
flask_cors
python-dotenv
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
